Title defaults to link when missing.  Blegh.
So... The previous commit was to get rid of html elements leaking into
the subject.  I tried lodash/underscore's unescape, but that only works
for the minimal amount of escaping to avoid XSS.  That means it did not
change &nbsp; to a space.  So, I tried node-ent, but that appeared to
replace &nbsp; with a utf-8 character, which made the (commented-out)
test fail with no visible differences, so I assume it's a difference of
utf8 vs non-utf8.