DB is migrated/cleared using knex
Oddly enough, it does not seem to be able to migrate the DB correctly in
the test.  When migrating via the spec helper, it does not record the
migration in the `knex_migrations` table, so tries to migrate it again
the next time around (which fails.)

And oh, past self, you are so cray-cray.  Once again, Mr. Frank was
right..  Utilizing the context object of the test runner is just a bad
idea.